# Conflicts:
#	server-part-service/src/main/java/com/example/server_part_service/controller/EntryController.java
#	server-part-service/src/main/java/com/example/server_part_service/dto/entry/RequestEntryDTO.java
#	server-part-service/src/main/java/com/example/server_part_service/dto/entry/ResponseEntryDTO.java
#	server-part-service/src/main/java/com/example/server_part_service/model/View.java
#	server-part-service/src/main/java/com/example/server_part_service/service/EntryService.java

+++ server-part-service/src/main/java/com/example/server_part_service/service/EntryService.java
@@ -3,7 +3,10 @@ package com.example.server_part_service.service;
 import com.example.server_part_service.dto.entry.RequestEntryDTO;
 import com.example.server_part_service.dto.entry.ResponseEntryDTO;
 import com.example.server_part_service.dto.entry.ResponseEntryDTOFourFields;
+import com.example.server_part_service.exception.AlreadyExistException;
+import com.example.server_part_service.exception.DeleteEntryException;
 import com.example.server_part_service.exception.EntityNotFoundException;
+import com.example.server_part_service.exception.EntryNotFoundException;
 import com.example.server_part_service.model.EntryModel;
 import com.example.server_part_service.model.ImageModel;
 import com.example.server_part_service.model.View;
@@ -24,6 +27,8 @@ public class EntryService {
     private EntryRepository entryRepository;
     @Autowired
     private ViewService viewService;
+    @Autowired
+    private ImageService imageService;
 
     public EntryModel getModelById(long entryId) {
         return entryRepository.findById(entryId)
@@ -36,24 +41,38 @@ public class EntryService {
     }
 
     public List<ResponseEntryDTO> findAll() {
-        List<EntryModel> entryModels = entryRepository.findAll();
-        List<ResponseEntryDTO> responseEntryDTOS = entryModels
-                .stream()
-                .map(this::converterModelToResponseDto) // Уберите круглые скобки
-                .toList(); // Преобразуем поток в список
-
-        return responseEntryDTOS; // Не забудьте вернуть список
+        try {
+            List<EntryModel> entryModels = entryRepository.findAll();
+            return entryModels
+                    .stream()
+                    .map(this::converterModelToResponseDto)
+                    .toList();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
+
     public ResponseEntryDTO getDtoById(long id) {
         return converterModelToResponseDto(getModelById(id));
     }
 
     public EntryModel saveModel(EntryModel entryModel) {
+        log.info("we in entryService 58");
+        if (existById(entryModel.getId())) {
+            throw new AlreadyExistException("already exist");
+        }
         return entryRepository.save(entryModel);
     }
 
+    private boolean existById(Long id) {
+        return entryRepository.existsById(id);
+    }
+
     public EntryModel updateModel(EntryModel entryModel) {
-        return entryRepository.save(entryModel);
+        if (existById(entryModel.getId())) {
+            return entryRepository.save(entryModel);
+        }
+        throw new EntryNotFoundException("error in update. reason not found");
     }
 
     public void deleteEntry(EntryModel model) {
@@ -61,10 +80,14 @@ public class EntryService {
     }
 
     public void deleteEntry(Long entryId) {
-        entryRepository.deleteById(entryId);
+        if (existById(entryId)) {
+            entryRepository.deleteById(entryId);
+        }
+        throw new DeleteEntryException("deleted error. reason: element not exist");
     }
 
     public ResponseEntryDTO converterModelToResponseDto(EntryModel model) {
+        log.info("we in en service 75");
         return new ResponseEntryDTO(
                 model.getId(),
                 model.getName(),
@@ -81,12 +104,15 @@ public class EntryService {
                 model.getNeededConservationActions(),
                 model.getSourcesOfInformation(),
                 model.getAuthors(),
-                getImageDataInBase64(model.getImageModel()),
+                model.getData().stream()
+                        .map(this::getImageDataInBase64)
+                        .collect(Collectors.toList()),
                 model.getView().getTitle()
         );
     }
 
     public EntryModel converterDtoToModel(RequestEntryDTO dto) {
+        log.info("we in entryService 99");
         return new EntryModel(
                 dto.getId(),
                 dto.getName(),
@@ -103,16 +129,21 @@ public class EntryService {
                 dto.getNeededConservationActions(),
                 dto.getSourcesOfInformation(),
                 dto.getAuthors(),
-                ImageService.convertDTOToImageModel(dto.getImage()),
-                new View(dto.getView()));
+                dto.getData().stream()
+                        .map(ImageService::convertDTOToImageModel)
+                        .collect(Collectors.toList()),
+                viewService.saveIfNotExist(new View(1L, dto.getView(),null))
+//                viewService.save(new View(0L, dto.getView(),null))
+        );
     }
 
-    public static String getImageDataInBase64(ImageModel image) {
+    public String getImageDataInBase64(ImageModel image) {
         log.info("{}", image);// Конвертация массива байтов в строку Base64
-        return "data:" + image.getContentType() + ";base64," + java.util.Base64.getEncoder().encodeToString(image.getData());}
+        return "data:" + image.getContentType() + ";base64," + java.util.Base64.getEncoder().encodeToString(image.getData());
+    }
 
 
- /*   public List<ResponseEntryDTOFourFields> findAllPreview() {
+    public List<ResponseEntryDTOFourFields> findAllPreview() {
         List<Object[]> fourFields1 = entryRepository.findFourFields();
         return fourFields1.stream()
                 .map(row -> {
@@ -120,11 +151,9 @@ public class EntryService {
                     dto.setId((Long) row[0]);
                     dto.setName((String) row[1]);
                     dto.setLatinName((String) row[2]);
-                    dto.setData(getImageDataInBase64((ImageModel) row[3])); // Преобразование data, предполагается, что это long
+                    dto.setData(getImageDataInBase64(imageService.findById((Long) row[3]))); // Преобразование data, предполагается, что это long
                     return dto;
                 })
                 .collect(Collectors.toList());
-
-                */
     }
-
+}
